requirements.txt
tensorflow>=2.15.0
fastapi
uvicorn
python-multipart
pillow
numpy